# docs/swagger.yaml
basePath: /
definitions:
  internal.ProductsSchema:
    properties:
      description:
        minLength: 20
        type: string
      image:
        type: string
      name:
        minLength: 3
        type: string
      price:
        type: number
    required:
    - description
    - name
    - price
    type: object
  internal.UserLoginSchema:
    properties:
      email:
        minLength: 6
        type: string
      password:
        minLength: 6
        type: string
    required:
    - password
    type: object
  internal.UserRegistrationSchema:
    properties:
      email:
        minLength: 6
        type: string
      name:
        minLength: 3
        type: string
      password:
        minLength: 6
        type: string
    required:
    - name
    - password
    type: object
  reponse.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  reponse.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:5050
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A complete e-commerce platform API built with Go, featuring user authentication,
    product management, payment processing with Stripe, and order management.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: E-Commerce API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password to get JWT token
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/internal.UserLoginSchema'
      produces:
      - application/json
      responses:
        "201":
          description: Login successful
          schema:
            $ref: '#/definitions/reponse.SuccessResponse'
        "400":
          description: Login failed
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
      summary: User login
      tags:
      - Authentication
  /order/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific order by its ID for the authenticated user
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order retrieved successfully
          schema:
            $ref: '#/definitions/reponse.SuccessResponse'
        "400":
          description: Bad request - Order ID required
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
        "401":
          description: Unauthorized - Authentication required
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - Orders
  /orders:
    get:
      consumes:
      - application/json
      description: Retrieve all orders for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Orders retrieved successfully
          schema:
            $ref: '#/definitions/reponse.SuccessResponse'
        "401":
          description: Unauthorized - Authentication required
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user orders
      tags:
      - Orders
  /payment/cancel:
    post:
      consumes:
      - application/json
      description: Handle cancelled payment callback from Stripe
      produces:
      - application/json
      responses:
        "200":
          description: Payment cancelled
          schema:
            $ref: '#/definitions/reponse.SuccessResponse'
      summary: Handle cancelled payment
      tags:
      - Payment
  /payment/checkout:
    post:
      consumes:
      - application/json
      description: Create a new Stripe checkout session for payment processing
      parameters:
      - description: Checkout session details
        in: body
        name: checkout
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Checkout session created successfully
          schema:
            $ref: '#/definitions/reponse.SuccessResponse'
        "400":
          description: Bad request - Invalid input data
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
        "401":
          description: Unauthorized - Authentication required
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Stripe checkout session
      tags:
      - Payment
  /payment/success:
    post:
      consumes:
      - application/json
      description: Handle successful payment callback from Stripe
      produces:
      - application/json
      responses:
        "200":
          description: Payment successful
          schema:
            $ref: '#/definitions/reponse.SuccessResponse'
        "400":
          description: Bad request - Missing session ID
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
      summary: Handle successful payment
      tags:
      - Payment
  /products:
    get:
      consumes:
      - application/json
      description: Retrieve all products with optional filtering and pagination
      parameters:
      - description: Filter by product name
        in: query
        name: name
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items to skip (default: 0)'
        in: query
        name: skip
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Products fetched successfully
          schema:
            $ref: '#/definitions/reponse.SuccessResponse'
        "400":
          description: Bad request - Invalid parameters
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
      summary: Get all products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product (requires authentication)
      parameters:
      - description: Product information
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/internal.ProductsSchema'
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            $ref: '#/definitions/reponse.SuccessResponse'
        "400":
          description: Bad request - Invalid input data
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
        "401":
          description: Unauthorized - Authentication required
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a product (requires authentication)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            $ref: '#/definitions/reponse.SuccessResponse'
        "400":
          description: Bad request - Product ID required
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
        "401":
          description: Unauthorized - Authentication required
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Retrieve a specific product by its ID (requires authentication)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product fetched successfully
          schema:
            $ref: '#/definitions/reponse.SuccessResponse'
        "400":
          description: Bad request - Product ID required
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
        "401":
          description: Unauthorized - Authentication required
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update an existing product (requires authentication)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated product information
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/internal.ProductsSchema'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            $ref: '#/definitions/reponse.SuccessResponse'
        "400":
          description: Bad request - Invalid input data or product ID
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
        "401":
          description: Unauthorized - Authentication required
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a product
      tags:
      - Products
  /registration:
    post:
      consumes:
      - application/json
      description: Register a new user account with name, email, and password
      parameters:
      - description: User registration information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal.UserRegistrationSchema'
      produces:
      - application/json
      responses:
        "201":
          description: Registration successful
          schema:
            $ref: '#/definitions/reponse.SuccessResponse'
        "400":
          description: Registration failed
          schema:
            $ref: '#/definitions/reponse.ErrorResponse'
      summary: Register a new user
      tags:
      - Authentication
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"